Always subtract closingtoken when getting doclength; misc tweaks.
PropertyValueInt is now actually a long (so it can accomodate token
lengths, which may well exceed 2 billion).

If the Document is already loaded, use that to get the document length
instead of DocValues.

+++ core/src/main/java/nl/inl/blacklab/resultproperty/PropertyValueDecade.java
@@ -22,7 +22,7 @@ public class PropertyValueDecade extends PropertyValueInt {
     public String toString() {
         if (value == HitPropertyDocumentDecade.UNKNOWN_VALUE)
             return "unknown";
-        int year = value - value % 10;
+        long year = value - value % 10;
         return year + "-" + (year + 9);
     }
 
@@ -30,6 +30,6 @@ public class PropertyValueDecade extends PropertyValueInt {
     public String serialize() {
         if (value == HitPropertyDocumentDecade.UNKNOWN_VALUE)
             return "unknown";
-        return PropertySerializeUtil.combineParts("dec", Integer.toString(value));
+        return PropertySerializeUtil.combineParts("dec", Long.toString(value));
     }
 }
